Remove factory.py
